+++ src/pacman/callback.h
@@ -33,7 +33,7 @@ void cb_trans_conv(pmtransconv_t event, void *data1, void *data2,
 
 /* callback to handle display of transaction progress */
 void cb_trans_progress(pmtransprog_t event, const char *pkgname, int percent,
-                   int howmany, int remain);
+                   size_t howmany, size_t remain);
 
 /* callback to handle receipt of total download value */
 void cb_dl_total(off_t total);
